Add major and minor version columns
Versions for collections look like this: <major_version>.<minor_version>

Every time a collection is updated, the major version will be
incremented.

Versions for modules look like this: <major_version>

Every time a module is updated, the major version will be incremented.
The collections that the module belongs to will all have their minor
versions incremented.

The original "version" column is not used anywhere in the cnx-archive
code now.  It is only there to store the legacy version.