Fix utf-8 search with no results
The problem is in post processing of the search results:

unmatched_terms = [term for term in query
                        if term[1] not in term_matches]

query is [('text', 'inf\xc3\xb8rmation')]
term_matches is [u'inf\xf8rmation']

Since one is bytes and one is unicode, it is considered not matched.
The fix is to make them both unicode.

Fix #188 and #222