Fix duplicate minor versions created by republish trigger
On legacy-cte-cnx-dev.cnx.org, when a user publishes a new version of a
page, the archive republish trigger should create a new minor version
for all the books that contain that page.  Instead of seeing one new
entry per book, there were two inserted into the modules table, with the
same minor version.

The reason is when we get the books that contain the page, the sql looks
in the trees table which may contain two trees for some books, one
collated, one not collated.  We end up getting the same book twice.

The solution is to use `SELECT DISTINCT` to make sure one book is only
republished once.