Make in-book-search work for collated/cooked content
Add an SQL method to determine if a particular book has a cooked
version. If so, use a different set of SQL queries for in-book (and
in-book-full-page) search. These differ because the table structure for
linking collated content to a book differs from that for the raw pages.
We may want to unify that, one day (context = NULL for uncooked?)