Fix multiple copies of new version after republish
We are seeing multiple copies of the same version in the modules table
after minor version updates.  Several pages are published one after the
other without waiting for the book to finish baking.  This results in
the same minor version being used for republish.

The reason for this seems to be because we use the latest_modules table
when we decide which books to republish.  The latest_modules table may
not actually have the latest versions of the books, because books only
end up in the latest_modules table once baking is finished.

There are 2 fixes to ensure this does not happen in the future:

1. look through the modules table for the max minor version so we would
   never add to the modules table with the same minor version,
2. use the modules table instead of the latest_modules table for
   deciding which books are latest and should be republished.

Close #508