Find nearest ancestor with id in xpath search
When user uses the xpath search, add `/ancestor-or-self::*[@id][1]` to
the end of their query to find nearest ancestor with an id so we can
generate a link to approximately that part of the document.

Change list link text to text without xml, and if there is no text,
display the tag.