Rewrite sitemap.py using lxml.etree
`sitemap.xml` was generated by hand, with an escape function in utils to
entity escape of urls.  Since `lxml.etree` is already used in other
parts of cnxarchive, we might as well make use of it and get rid of the
escape function in utils.  This will also make sure the sitemap.xml is
valid xml.