Set timeout on travis waiting for archive container to be up
A check was added to .travis.yml for waiting for the archive container
to be available, but if the archive container exited for whatever
reason, this loop will run forever and not ever timeout.  We don't
really expect the container to take that long to be available, so the
timeout is set to ~1 minute.