Add a jenkinsfile for releasing the container and package
This builds the container and publishes it to the container registry.

This also builds the Python distribution/package and uploads it to pypi.org. Though, at this time it is configured to work against the test instance of pypi.org. Once we've verified that works, we can remove those lines.