Adjust the container build steps (#616)
* Adjust the container build steps

This moves away from the installing the software from a foreign requirements.txt file. The requirements.txt was not installing the software in the current working directory, but rather the contents of a git checkout for this repo. Then it was installing from source.

Instead these changes now install from the source checked out on the user's machine. It also installs released dependency packages.

* Install coverage tools system wide

* Assign ownership of the container's source directory to the runtime user

* Fix container source directory permissions