Fix the recent.rss feed (#617)
* Fix the recent.rss feed

For starters, it was invalid xml, so that's fixed. =)

I've added some of the optional rss v2.0 tags to allow readers to derive uniqueness of the items, since many of the results have the same title (i.e. Introduction).

* Fix to remove trailing newline

My editor insists on doing the correct thing... putting in a newline char at the end of the file.