Fix the package release stage using generic container (#623)
Changing the Dockerfile to run the container as an unprivilaged user
has made the existing release stage command fail due to permissions errors
while installing twine. To keep it a single line command, I'm simply
switching to use the base `python:2.7` container. The results will be the same
because the `MANIFEST.in` dictates what goes in a package,
not the container's contents.