Output container log if test fails to start on travis
It used to be that travis would fail after the maximum time limit for
the job when the archive container exited.  This is a bit inconvenient
because the user might think the tests are just taking a long time.

We have a loop that waits for the archive container to start before
running the tests. If the archive container exits, we should print out
the logs and not continue waiting for it to be up.