Mount files in docker compose separately due to permissions
The archive docker container uses user "archive" which does not have
permission to write to the `/src/` directory as it is owned by root
after mounting the host directory.

This is the error when running `docker-compose run --rm archive /bin/bash`:

```
archive@96e0dc78b676:/src$ python setup.py test
running test
error: can't create or remove files in install directory

The following error occurred while trying to add or remove files in the
installation directory:

    [Errno 13] Permission denied: './.eggs/test-easy-install-9.write-test'
```

It appears there is currently no way of mounting a host directory as a
user.  Doing `chown` in the container messes up the owner host directory
so mounting the files separately is the only solution I came up with.