Fix collection.xml comparison in collection minor update test
In the collection minor update test, we are comparing the new
collection.xml generated in the database with the one in test/data.  The
code tries to remove the revised timestamp by doing:

```
            rev_offset = result.find('<md:revised>')
            rev_start = rev_offset + 12  # len('<md:revised>')
            rev_end = rev_start + 32  # len(timestamp)
            self.assertEqual(result[:rev_start]+result[rev_end:],
                             expected[:rev_start]+expected[rev_end:])
```

This fails sometimes because timestamp appears to be not always 32
characters long.  For example, one time it got truncated to
`<md:revised>/md:revised>`.

Avoid this issue by using regular expression to remove everything inside
the `<md:revised>` tag.