XPath search with support for searching CNXML, HTML and Baked HTML content (#635)
* Extract params to later DRY things

This extracts the query string parameters using a decorator. The decorator
will be used in a followup commit to DRY the function when the route is split
as json and html. This makes the request response proceedure in the function
clearly read.

* Cut to the chase (rm and rewrite)

* Rewrote base query functions for xpath search

This set of functions first gathers the contextual set of documents
to query, then does the xpath query on them. As opposed to doing all
that in a single sql query, which makes it very difficult to read.

* Correct the xpath search view tests

This still doesn't address the rendering of the response,
but it does not address getting results from the view.

* Add the xpath search json and html renderings

* pep8 fixes

* Add the ability to xpath search by html and baked-html

* Remove inline development SQL

* Fix utf-8 chars in response data