Revert "handle sitemap.xml with no authorname (from_id)"
This reverts commit 8ef2c1573ee4a98224a7cd698d050c3ba273fd5a.

The primary reason we are reverting this is because it allows
for bad things to happen.

Also, it's not clear why this view function needs to be able to handle
a value that must be passed in from the framework's perspective.
We tie this to a route that requires something be given as the url
parameter, so it's never going to be null.